rename flatex.py to flatex_core.py
