+++ conda/meta.yaml
@@ -14,7 +14,7 @@ build:
   number: 0
   script: {{ PYTHON }} -m pip install . -vv
   entry_points:
-    - flatexpy = flatexpy.flatexpy:main
+    - flatexpy = flatexpy.flatexpy_core:main
 
 requirements:
   host:
