+++ tests/integration/test_configuration.py
@@ -3,7 +3,7 @@
 import os
 import tempfile
 
-from flatexpy.flatexpy import LatexExpandConfig, LatexExpander
+from flatexpy.flatexpy_core import LatexExpandConfig, LatexExpander
 
 
 class TestConfigurationOptions:
