+++ tests/integration/test_error_handling.py
@@ -5,7 +5,7 @@ import tempfile
 
 import pytest
 
-from flatexpy.flatexpy import LatexExpandConfig, LatexExpander, LatexExpandError
+from flatexpy.flatexpy_core import LatexExpandConfig, LatexExpander, LatexExpandError
 
 
 class TestErrorHandling:
