+++ tests/test_integration.py
@@ -10,7 +10,7 @@ import tempfile
 
 import pytest
 
-from flatexpy.flatexpy import LatexExpandConfig, LatexExpander, LatexExpandError
+from flatexpy.flatexpy_core import LatexExpandConfig, LatexExpander, LatexExpandError
 
 
 class TestBasicIntegration:
