+++ tests/unit/test_expander_core.py
@@ -7,7 +7,7 @@ from unittest.mock import MagicMock, mock_open, patch
 
 import pytest
 
-from flatexpy.flatexpy import (
+from flatexpy.flatexpy_core import (
     GraphicsNotFoundError,
     LatexExpandConfig,
     LatexExpander,
