+++ tests/unit/test_file_processing.py
@@ -7,7 +7,7 @@ from unittest.mock import MagicMock, mock_open, patch
 
 import pytest
 
-from flatexpy.flatexpy import LatexExpandConfig, LatexExpander, LatexExpandError
+from flatexpy.flatexpy_core import LatexExpandConfig, LatexExpander, LatexExpandError
 
 
 class TestFileProcessing:
